requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
anyio==4.0.0
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
arviz==0.17.1
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
Babel==2.14.0
beautifulsoup4==4.12.2
bertopic==0.15.0
bleach==6.1.0
bokeh==3.2.2
cachetools==5.3.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
cloudpickle==2.2.1
colorcet==3.0.1
comm==0.2.2
cons==0.4.6
contourpy==1.1.1
cycler==0.12.1
Cython==0.29.37
dask==2023.9.2
dask-cuda==23.10.0
datashader==0.15.2
datashape==0.5.2
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.8
distributed==2023.9.2
etuples==0.3.9
exceptiongroup==1.1.3
executing==2.0.1
fastjsonschema==2.18.1
fastprogress==1.0.3
fastrlock==0.8.2
filelock==3.13.1
fiona==1.9.5
flatbuffers==23.5.26
fonttools==4.43.1
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.9.2
gast==0.5.4
gensim==4.3.2
geopandas==0.14.0
google-auth==2.23.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
graphviz==0.20.1
grpcio==1.59.0
h11==0.14.0
h5netcdf==1.3.0
h5py==3.10.0
hdbscan==0.8.33
holoviews==1.17.1
httpcore==1.0.4
httpx==0.27.0
huggingface-hub==0.17.3
idna==3.6
importlib-metadata==6.8.0
ipykernel==6.29.3
ipython==8.22.2
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
json5==0.9.22
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter-events==0.8.0
jupyter-lsp==2.2.4
jupyter_client==8.4.0
jupyter_core==5.4.0
jupyter_server==2.8.0
jupyter_server_proxy==4.1.0
jupyter_server_terminals==0.4.4
jupyterlab==4.1.4
jupyterlab-pygments==0.2.2
jupyterlab_server==2.25.4
keras==2.14.0
kiwisolver==1.4.5
lazy_loader==0.3
libclang==16.0.6
linkify-it-py==2.0.2
llvmlite==0.40.1
locket==1.0.0
logical-unification==0.4.6
Markdown==3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.0
matplotlib-inline==0.1.6
mdit-py-plugins==0.4.0
mdurl==0.1.2
miniKanren==1.0.3
mistune==3.0.2
ml-dtypes==0.2.0
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.4
multipledispatch==1.0.0
nbclient==0.8.0
nbconvert==7.9.2
nbformat==5.9.2
nest-asyncio==1.6.0
networkx==3.1
nltk==3.8.1
notebook_shim==0.2.4
numba==0.57.1
numpy==1.24.4
oauthlib==3.2.2
opt-einsum==3.3.0
orjson==3.9.9
overrides==7.4.0
packaging==23.2
pandas==1.5.3
pandocfilters==1.5.0
panel==1.2.3
param==1.13.0
parso==0.8.3
partd==1.4.1
pexpect==4.9.0
Pillow==10.1.0
platformdirs==3.11.0
plotly==5.17.0
prometheus-client==0.17.1
prompt-toolkit==3.0.43
protobuf==4.24.4
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==12.0.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pyct==0.5.0
Pygments==2.17.2
pymc==5.10.4
pynndescent==0.5.10
pynvml==11.4.1
pyparsing==3.1.2
pyproj==3.6.1
pytensor==2.18.6
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
pytz==2024.1
pyviz_comms==3.0.0
PyYAML==6.0.1
pyzmq==25.1.1
referencing==0.30.2
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.10.6
rsa==4.9
safetensors==0.4.2
scikit-learn==1.3.1
scipy==1.11.3
Send2Trash==1.8.2
sentence-transformers==2.2.2
sentencepiece==0.1.99
shapely==2.0.2
simpervisor==1.0.0
six==1.16.0
smart-open==6.4.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
stack-data==0.6.3
sympy==1.12
tblib==2.0.0
tenacity==8.2.3
tensorboard==2.14.1
tensorboard-data-server==0.7.1
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-hub==0.15.0
tensorflow-io-gcs-filesystem==0.34.0
tensorflow-macos==2.14.0
termcolor==2.3.0
terminado==0.17.1
threadpoolctl==3.3.0
tinycss2==1.2.1
tokenizers==0.14.1
tomli==2.0.1
toolz==0.12.0
top2vec==1.0.29
torch==2.1.0
torchaudio==2.1.0
torchvision==0.16.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.14.2
transformers==4.34.0
treelite==3.9.1
treelite-runtime==3.9.1
types-python-dateutil==2.8.19.14
typing_extensions==4.10.0
tzdata==2023.3
uc-micro-py==1.0.2
umap-learn==0.5.4
uri-template==1.3.0
urllib3==2.2.1
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
Werkzeug==3.0.0
wordcloud==1.9.2
wrapt==1.14.1
xarray==2023.9.0
xarray-einstats==0.7.0
xyzservices==2023.10.0
yarl==1.9.2
zict==3.0.0
zipp==3.18.0
